+++ CHANGES.rst
@@ -6,6 +6,8 @@ Unreleased
 
 * ControlMessage and ControlRequest are now simply Message and Request,
   as there is no other kind of message or request.
+* The ``search`` module is now ``files``, and so is the corresponding
+  DVRIPClient method.
 
 0.0.3 (2019-05-03)
 ------------------